add update feture

+++ src/Redux-thunk/Crud/View.jsx
@@ -1,6 +1,7 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import {  deleteUser, ViewUser } from "../UserSlice";
+import { deleteUser, ViewUser } from "../UserSlice";
+import { NavLink } from "react-router-dom";
 
 const View = () => {
   const { userList } = useSelector((state) => state.users);
@@ -10,44 +11,55 @@ const View = () => {
     dispatch(ViewUser());
   }, [dispatch]);
 
-  function trash(id){
-    dispatch(deleteUser(id))
-    // console.log(id);
+  function trash(id) {
+    if(confirm("Do you Want to delete this item ?")){
+
+      dispatch(deleteUser(id));
+      console.log(`Deleted User ID: ${id}`);
+    }
   }
-  // console.log(userList.id);
-  
 
   return (
     <>
       <div className="container my-5">
-        <h3 className="text-center mt-5">UserList</h3>
-
-<div className="table-responsive">
-        <table className="table-bordered table-striped rounded col-12 table-hover   ">
-          <thead className="bg-dark text-white">
-            <tr>
-              <th>SNO</th>
-              <th>UserName</th>
-              <th>Date</th>
-              <th>Decription</th>
-              <th>Action</th>
-            </tr>
-          </thead>
-          <tbody>
-            {userList.map((user, index) => (
+        <h3 className="text-center mt-5">User List</h3>
+        <div className="table-responsive">
+          <table className="table-bordered table-striped rounded col-12 table-hover">
+            <thead className="bg-dark text-white">
               <tr>
-                <td>{index + 1}</td>
-                <td>{user.name}</td>
-                <td>{user.blog}</td>
-                <td>{user.des}</td>
-                <td>
-                  <button className="btn btn-danger" onClick={()=>trash(user.id)}>Delete</button>
-                </td>
+                <th>SNO</th>
+                <th>UserName</th>
+                <th>Date</th>
+                <th>Description</th>
+                <th>Action</th>
               </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
+            </thead>
+            <tbody>
+              {userList.map((user, index) => (
+                <tr key={user.id}>
+                  <td>{index + 1}</td>
+                  <td>{user.name}</td>
+                  <td>{user.blog}</td>
+                  <td>{user.des}</td>
+                  <td>
+                    <button
+                      className="btn btn-danger btn-sm"
+                      onClick={() => trash(user.id)}
+                    >
+                      Delete
+                    </button>
+                    <NavLink
+                      className="btn btn-secondary btn-sm"
+                      to={`/Update/${user.id}`}
+                    >
+                      Update
+                    </NavLink>
+                  </td>
+                </tr>
+              ))}
+            </tbody>
+          </table>
+        </div>
       </div>
     </>
   );
